subgraphs/events/requirements.txt
fastapi
strawberry-graphql[fastapi]
uvicorn
motor